Move memory system configs to a separate configs file
